[flink] Refactor code in RewritePostponeBucketCommittableOperator
